+++ includes/utils.h
@@ -3,8 +3,8 @@
 
 #include <stdlib.h>
 
-extern void afficher_tableau(double *array, size_t n);
-extern void afficher_tableau_trig_double(double *array, size_t n);
+extern void afficher_tableau(long double *array, size_t n);
+extern void afficher_tableau_trig_double(long double *array, size_t n);
 extern void afficher_tableau_trig_int(int *array, size_t n);
 
 #endif
